fix: move dir
